\name{simecolUtil}
\alias{simecolUtil-package}
\alias{simecolUtil}
\docType{package}
\title{
  simecolUtil - A Collection of Potentially Helpful Tools
}
\description{
  This collection is a temporary storage for functions of different quality, 
  that may eventually be transferred to other packages.
}
\details{
\tabular{ll}{
Package: \tab simecolUtil\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2011-01-29\cr
License: \tab GPL (>= 2) \cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
Thomas Petzoldt <thomas.petzoldt@tu-dresden.de>
}
%\references{
%~~ Literature or other references for background information ~~
%}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
