\name{tProfile}
\alias{tProfile}
\alias{t_xyz}
\alias{t_depth}
\title{Temperature Profile Data of Bautzen Reservoir}

\description{The dataset is a one year subset of water temperature data, 
 measured during 1995 in a shallow reservoir.
}

\format{t_depth is a data frame with the first column giving the Julian 
        day of the measurement and 11 columns giving the water temperature of 
        each depth from surface (0 m) up to 10 m below surface.

\code{t_xyz} contains the same data, but as three columns day, depth, temp.	
}

\usage{
data(t_depth)

data(t_xyz)
}

\source{Dresden University of Technology, Institute of Hydrobiology}

\keyword{datasets}
