\name{tsbshore}
\alias{tsbshore}

\title{Shore Line of Bautzen Reservoir}

\description{The data set gives the coordinates (in m) of the shoreline 
             of Bautzen Reservoir (Saxony, Germany).
}

\usage{data(tsbshore)}


\format{x and y coordinates}

\source{TU Dresden, Institute of Hydrobiology}

\examples{
  data(tsbshore)
  plot(tsbshore, type = "l")
}

\keyword{datasets}
